Move socket implementation into module http_util
This removes the requirement of linking socket libraries for
applications that do not use http_util
